chore: refactor overlay component according to new state structure
