+++ examples/basic/src/components/index.ts
@@ -3,5 +3,6 @@ export * from './Seeker';
 export * from './AudioTracksSelector';
 export * from './VideoTracksSelector';
 export * from './TextTracksSelector';
+export * from './Overlay';
 export * from './TopControl';
 export * from './Toast';
